Renaming input to Input requires this middle step
